AppConfig: Move default for rendererOrigin to config file
As we only use rendererOrigin in the frontend config service, where domain will be used if it is not defined, it makes more sense to move this default behaviour to the app config directly. That makes it easier to understand what this variable contains and that it default to domain.

Signed-off-by: Philip Molares <philip.molares@udo.edu>